Working towards getting stuff rendered again! :tada:
